+++ packages/framer-motion/src/animation/animators/MainThreadAnimation.ts
@@ -109,6 +109,7 @@ export class MainThreadAnimation<
         super(options)
 
         const { name, motionValue, keyframes } = this.options
+
         const onResolved = (
             resolvedKeyframes: ResolvedKeyframes<T>,
             finalKeyframe: T
@@ -225,6 +226,14 @@ export class MainThreadAnimation<
     }
 
     tick(timestamp: number, sample = false) {
+        const { resolved } = this
+
+        // If the animations has failed to resolve, return the final keyframe.
+        if (!resolved) {
+            const { keyframes } = this.options
+            return { done: true, value: keyframes[keyframes.length - 1] }
+        }
+
         const {
             finalKeyframe,
             generator,
@@ -234,7 +243,7 @@ export class MainThreadAnimation<
             calculatedDuration,
             totalDuration,
             resolvedDuration,
-        } = this.resolved
+        } = resolved
 
         if (this.startTime === null) return generator.next(0)
 
@@ -387,7 +396,8 @@ export class MainThreadAnimation<
     state: AnimationPlayState = "idle"
 
     get duration() {
-        return millisecondsToSeconds(this.resolved.calculatedDuration)
+        const { resolved } = this
+        return resolved ? millisecondsToSeconds(resolved.calculatedDuration) : 0
     }
 
     get time() {
@@ -472,13 +482,13 @@ export class MainThreadAnimation<
     }
 
     stop() {
+        this.resolver.cancel()
         this.isStopped = true
         if (this.state === "idle") return
 
-        this.state = "idle"
+        this.teardown()
         const { onStop } = this.options
         onStop && onStop()
-        this.teardown()
     }
 
     complete() {
@@ -503,6 +513,7 @@ export class MainThreadAnimation<
             this.tick(this.cancelTime)
         }
         this.teardown()
+        this.updateFinishedPromise()
     }
 
     private teardown() {