+++ packages/framer-motion/src/projection/node/types.ts
@@ -98,11 +98,11 @@ export interface IProjectionNode<I = unknown> {
 
     setTargetDelta(delta: Delta): void
     resetTransform(): void
-    resetRotation(): void
+    resetSkewAndRotation(): void
     applyTransform(box: Box, transformOnly?: boolean): Box
     resolveTargetDelta(force?: boolean): void
     calcProjection(): void
-    getProjectionStyles(styles?: MotionStyle): MotionStyle | undefined
+    getProjectionStyles(styleProp?: MotionStyle): MotionStyle | undefined
     clearMeasurements(): void
     resetTree(): void